% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_brick.R
\name{aggregate_brick}
\alias{aggregate_brick}
\title{Parallel aggregate}
\usage{
aggregate_brick(dummie_nc, new_res)
}
\arguments{
\item{dummie_nc}{a character string}

\item{new_res}{numeric}
}
\value{
raster brick
}
\description{
Function to aggregate a raster brick
}
\keyword{internal}
