% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_info.R
\name{aux_info}
\alias{aux_info}
\title{Complement fldmean}
\usage{
aux_info(dummie_name)
}
\arguments{
\item{dummie_name}{character}
}
\value{
character
}
\description{
Function to add name and type
}
\keyword{internal}
