% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_ts.R
\name{aux_ts}
\alias{aux_ts}
\title{Complement make_ts generated csv file}
\usage{
aux_ts(dummie_name)
}
\arguments{
\item{dummie_name}{character}
}
\value{
character
}
\description{
Function to generate name and type columns
}
\keyword{internal}
