% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick_to_dt.R
\name{brick_to_dt}
\alias{brick_to_dt}
\title{Transform raster into data.table}
\usage{
brick_to_dt(x)
}
\arguments{
\item{x}{a character string with the path to the data file.}
}
\value{
data.table
}
\description{
Function to transform a raster brick into a data.table
}
\examples{
\dontrun{
x <- brick_to_dt("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
y <- brick_to_dt("dummie.nc")
}
}
