% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick_trends.R
\name{brick_trends}
\alias{brick_trends}
\title{Trends}
\usage{
brick_trends(x, annual = NULL)
}
\arguments{
\item{x}{a character string with the path to the data file. Or a RasterBrick.}

\item{annual}{a character string with the desired aggregation function. Suitable options are:
\itemize{
\item "max"
\item "mean"
\item "median"
\item "min"
\item "sum"
}}
}
\value{
a RasterLayer with trend values
}
\description{
Function to compute linear slope of each raster grid cell
}
