% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_csv.R
\name{cdo_csv}
\alias{cdo_csv}
\title{Format cdo generated csv file}
\usage{
cdo_csv(dummie_nc)
}
\arguments{
\item{dummie_nc}{a character string}
}
\value{
No return value, called to save a csv file
}
\description{
Function to compute the weighted average of an .nc file
}
\keyword{internal}
