% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Directory creator}
\usage{
create_folders(destination)
}
\arguments{
\item{destination}{a character string with the path where the folders will be created as follows:
\itemize{
\item{\code{destination}/data}
\itemize{
\item{\code{destination}/data/database}
\item{\code{destination}/data/processed}
}
}}
}
\value{
No return value, called to create tidy directories for pRecipe.
}
\description{
Function for creating tidy folders to store pRecipe.
}
\keyword{internal}
