% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R
\name{csi}
\alias{csi}
\title{Probability of Detection}
\usage{
csi(x, ref, th)
}
\arguments{
\item{x}{a data.table generated by \code{\link{fldmean}}}

\item{ref}{a data.table with data used for evaluation}

\item{th}{numeric. The value for detection threshold}
}
\value{
numeric
}
\description{
Function for calculating the critical success index.
}
