% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_check.R
\name{dataset_check}
\alias{dataset_check}
\title{Data set name checker}
\usage{
dataset_check(data_name)
}
\arguments{
\item{data_name}{a character string.}
}
\value{
No return value, called to download the data set.
}
\description{
Function to check if the data set is available
}
\keyword{internal}
