% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{display_data}
\alias{display_data}
\title{Display available data sets}
\usage{
display_data()
}
\value{
a data.frame listing the data sets pRecipe is able to work with.
}
\description{
Function for displaying available data sets in pRecipe.
}
