% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_20cr.R
\name{download_20cr}
\alias{download_20cr}
\title{20CR data downloader}
\usage{
download_20cr(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading 20CR v3.
}
\keyword{internal}
