% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cmap.R
\name{download_cmap}
\alias{download_cmap}
\title{CMAP data downloader}
\usage{
download_cmap(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CMAP standard version.
}
\keyword{internal}
