% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cmorph.R
\name{download_cmorph}
\alias{download_cmorph}
\title{CMORPH data downloader}
\usage{
download_cmorph(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CMORPH.
}
\keyword{internal}
