% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_cru}
\alias{download_cru}
\title{CRU data downloader}
\usage{
download_cru(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading CRU_TS NC.GZ file.
}
