% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cru_ts.R
\name{download_cru_ts}
\alias{download_cru_ts}
\title{CRU_TS data downloader}
\usage{
download_cru_ts(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CRU_TS v4.06.
}
\keyword{internal}
