% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_e_obs.R
\name{download_e_obs}
\alias{download_e_obs}
\title{E-OBS data downloader}
\usage{
download_e_obs(folder_path = ".", time_res = "monthly")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}

\item{time_res}{a character string with the desired time resolution. Suitable options are:
\itemize{
\item{"monthly",}
\item{"yearly".}
}}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading E-OBS.
}
