% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_em_earth.R
\name{download_em_earth}
\alias{download_em_earth}
\title{EM_EARTH data downloader}
\usage{
download_em_earth(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading EM-EARTH.
}
\keyword{internal}
