% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era20c.R
\name{download_era20c}
\alias{download_era20c}
\title{ERA20C data downloader}
\usage{
download_era20c(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading ERA-20C.
}
\keyword{internal}
