% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era5.R
\name{download_era5}
\alias{download_era5}
\title{ERA5 data downloader}
\usage{
download_era5(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading ERA5.
}
\keyword{internal}
