% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcn.R
\name{download_ghcn}
\alias{download_ghcn}
\title{GHCN data downloader}
\usage{
download_ghcn(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GHCN v2.
}
\keyword{internal}
