% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpcc.R
\name{download_gpcc}
\alias{download_gpcc}
\title{GPCC data downloader}
\usage{
download_gpcc(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GPCC v2020.
}
\keyword{internal}
