% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_gpcp}
\alias{download_gpcp}
\title{GPCP data downloader}
\usage{
download_gpcp(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading GPCP NC file.
}
