% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_gpm}
\alias{download_gpm}
\title{GPM data downloader}
\usage{
download_gpm(destination, start_year, end_year)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}

\item{start_year}{numeric. Start year should be between 2000-2019.}

\item{end_year}{numeric. End year should be between 2000-2019, and should be greater or equal to start year.}
}
\description{
Function for downloading GPM HDF5 files.
}
