% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_gpm_imergm}
\alias{download_gpm_imergm}
\title{GPM data downloader}
\usage{
download_gpm_imergm(destination, start_year = 2000, end_year = 2020)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}

\item{start_year}{numeric. Start year should be between 2000-2020.}

\item{end_year}{numeric. End year should be between 2000-2020, and should be greater or equal to start year.}
}
\description{
Function for downloading GPM HDF5 files.
}
\note{
user must \href{"https://urs.earthdata.nasa.gov"}{Create an Earthdata account} and \href{https://disc.gsfc.nasa.gov/earthdata-login}{Link GES DISC}
}
