% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_ncep}
\alias{download_ncep}
\title{NCEP/NCAR data downloader}
\usage{
download_ncep(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading NCEP/NCAR NC files.
}
