% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_ncep2}
\alias{download_ncep2}
\title{NCEP/DOE data downloader}
\usage{
download_ncep2(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading NCEP/DOE NC files.
}
