% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_ncep_ncar}
\alias{download_ncep_ncar}
\title{NCEP/NCAR data downloader}
\usage{
download_ncep_ncar(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading NCEP/NCAR NC files.
}
