% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_precl.R
\name{download_precl}
\alias{download_precl}
\title{PRECL data downloader}
\usage{
download_precl(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading PREC/L.
}
\keyword{internal}
