% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_trmm_3b43}
\alias{download_trmm_3b43}
\title{TRMM data downloader}
\usage{
download_trmm_3b43(folder_path, start_year = 1998, end_year = 2019)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}

\item{start_year}{numeric. Start year should be between 1998-2019.}

\item{end_year}{numeric. End year should be between 1979-2019, and should be greater or equal to start year.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading TRMM 3B43 HDF files.
}
\note{
user must \href{https://urs.earthdata.nasa.gov}{Create an Earthdata account} and \href{https://disc.gsfc.nasa.gov/earthdata-login}{Link GES DISC}
}
