% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_udel.R
\name{download_udel}
\alias{download_udel}
\title{UDEL data downloader}
\usage{
download_udel(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading UDEL v5.01.
}
\keyword{internal}
