% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dt_aggregate}
\alias{dt_aggregate}
\title{Data table aggregation in space}
\usage{
dt_aggregate(data, res)
}
\arguments{
\item{data}{data.table. A precipitation data table reformatted by pRecipe.}

\item{res}{numeric. The upscale target resolution.}
}
\value{
the aggregated data table at the new spatial resolution.
}
\description{
Function for upscaling spatial resolution on pRecipe data tables
}
