% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dt_parallel}
\alias{dt_parallel}
\title{Data table weighted mean for parallel computing}
\usage{
dt_parallel(dummie_table)
}
\arguments{
\item{dummie_table}{A pRecipe data table with multiple data sets.}
}
\value{
a data table with the weighted average of the original.
}
\description{
Function for merging data sets available in pRecipe data table by weighted mean.
}
