% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_ce_ts.R
\docType{data}
\name{era5_ce_ts}
\alias{era5_ce_ts}
\title{Monthly precipitation data}
\format{
A data.table with 480 obs. of 2 variables:
\describe{
  \item{date}{IDate format \%Y-\%m-\%d}
  \item{value}{monthly average values}
}
}
\source{
European Centre for Medium-Range Weather Forecasts (ECMWF)
}
\usage{
era5_ce_ts
}
\description{
A subset of ERA5 monthly precipitation data in mm over 2-28E, 42-58N. More details of the raw data can be found \href{https://www.ecmwf.int/en/forecasts/datasets/reanalysis-datasets/era5}{here}.
}
\keyword{datasets}
