% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_check.R
\name{exists_check}
\alias{exists_check}
\title{Existance checker}
\usage{
exists_check(data_name)
}
\arguments{
\item{data_name}{a character string.}
}
\value{
No return value, called to download the data set.
}
\description{
Function to check if the data file already exists.
}
\keyword{internal}
