% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far.R
\name{far}
\alias{far}
\title{False Alarm Rate}
\usage{
far(x, ref, th)
}
\arguments{
\item{x}{a data.table generated by \code{\link{make_ts}}}

\item{ref}{a data.table with data used for evaluation}

\item{th}{numeric. The value for detection threshold}
}
\value{
numeric
}
\description{
Function for calculating the false alarm rate.
}
