% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_name_out.R
\name{fix_name_out}
\alias{fix_name_out}
\title{Fix name out}
\usage{
fix_name_out(nc_out)
}
\arguments{
\item{nc_out}{a character string}
}
\value{
No return value, called to rename subsetted file
}
\description{
Function to fix the names when subsetting
}
\keyword{internal}
