% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_cz_ts.R
\docType{data}
\name{gpm_cz_ts}
\alias{gpm_cz_ts}
\title{Monthly precipitation data}
\format{
A data.table with 120 obs. of 4 variables:
\describe{
  \item{date}{IDate format \%Y-\%m-\%d}
  \item{value}{monthly average values}
  \item{name}{full name of the data set}
  \item{type}{source type of the data set}
}
}
\source{
National Aeronautics and Space Administration (NASA)
}
\usage{
gpm_cz_ts
}
\description{
A subset of GPM-IMERGM monthly precipitation data in mm over Czechia. More details of the raw data can be found \href{https://gpm.nasa.gov/data/imerg}{here}.
}
\keyword{datasets}
