% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_subset_ts.R
\docType{data}
\name{gpm_subset_ts}
\alias{gpm_subset_ts}
\title{Monthly precipitation data}
\format{
A data.table with 180 obs. of 2 variables:
\describe{
  \item{date}{Date format \%Y-\%m-\%d}
  \item{value}{monthly average values}
}
}
\source{
National Aeronautics and Space Administration (NASA)
}
\usage{
gpm_subset_ts
}
\description{
A subset of GPM-IMERGM monthly precipitation data in mm over South America. More details of the raw data can be found \href{https://gpm.nasa.gov/data/imerg}{here}.
}
\keyword{datasets}
