% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_cpc}
\alias{import_cpc}
\title{CPC data reader}
\usage{
import_cpc(folder_path, chk = FALSE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a list of bricks with daily precipitation data in [mm] and missing values are NA.
}
\description{
Function for reading CPC-GLOBAL NC files.
}
