% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_gpcp}
\alias{import_gpcp}
\title{GPCP data reader}
\usage{
import_gpcp(file_path)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}
}
\description{
Function for reading GPCP NC file.
}
