% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_gpm}
\alias{import_gpm}
\title{GPM data reader}
\usage{
import_gpm(folder_path, chk = FALSE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a list of bricks with monthly precipitation data in [mm/day] and missing values are NA.
}
\description{
Function for reading GPM HDF5 files.
}
