% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_precl}
\alias{import_precl}
\title{PRECL data reader}
\usage{
import_precl(file_path, resolution)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}

\item{resolution}{numeric. Data spatial resolution. Suitable options are:
\itemize{
\item{0.5 for 0.5 degree,}
\item{1 for 1 degree,}
\item{2.5 for 2.5 degree.}
}}
}
\description{
Function for reading PRECL NC file.
}
