% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_udel}
\alias{import_udel}
\title{UDEL data reader}
\usage{
import_udel(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a brick with monthly precipitation data in [mm] and missing values are NA.
}
\description{
Function for reading UDEL NC file.
}
