% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ts.R
\name{make_ts}
\alias{make_ts}
\title{Generate time series}
\usage{
make_ts(data, autosave = FALSE)
}
\arguments{
\item{data}{a character string with the path to the data file. Or a RasterBrick}

\item{autosave}{logical FALSE (default). If TRUE data will be automatically stored in the same location of the input file}
}
\value{
A data.table
}
\description{
The function \code{make_ts} generates a csv time series and stored in the same location of the input file.
}
\examples{
\dontrun{
make_ts("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
make_ts("dummie.nc")
}
}
