% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_1891_1899}
\alias{merge_1891_1899}
\title{Data integrator for 1891_1899.}
\usage{
merge_1891_1899(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "integration" folder.}
}
\description{
Function for merging the data sets overlapping by time periods.
}
