% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_1891_2019}
\alias{merge_1891_2019}
\title{Data integrator for 1891_2019.}
\usage{
merge_1891_2019(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "database" folder.}
}
\description{
Function for merging the data sets overlapping by time periods.
}
