% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_1891_2020}
\alias{merge_1891_2020}
\title{Data integrator for 1891_2020.}
\usage{
merge_1891_2020(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "database" folder.}
}
\value{
No return value, called for side effects.
}
\description{
Function for merging the data sets overlapping by time periods.
}
