% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_1901_1947}
\alias{merge_1901_1947}
\title{Data integrator for 1901_1947.}
\usage{
merge_1901_1947(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "integration" folder.}
}
\description{
Function for merging the data sets overlapping by time periods.
}
