% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_1948_1978}
\alias{merge_1948_1978}
\title{Data integrator for 1948_1978.}
\usage{
merge_1948_1978(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "integration" folder.}
}
\description{
Function for merging the data sets overlapping by time periods.
}
