% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_1998_2000}
\alias{merge_1998_2000}
\title{Data integrator for 1998_2000}
\usage{
merge_1998_2000(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "integration" folder.}
}
\description{
Function for merging the data sets overlapping by time periods.
}
