% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_cdr}
\alias{merge_cdr}
\title{Climate Data Record}
\usage{
merge_cdr(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "integration" folder.}
}
\description{
Function for unifying the merged periods.
}
