% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{merge_time}
\alias{merge_time}
\title{Data integrator by period.}
\usage{
merge_time(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "integration" folder.}
}
\description{
Function for merging the data sets overlaping by time periods.
}
