% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mon_to_year.R
\name{mon_to_year}
\alias{mon_to_year}
\title{Rescale a precipitation data product in time}
\usage{
mon_to_year(name, database_path = "./data/database")
}
\arguments{
\item{name}{a Raster object or a character string with the name(s) of the desired data set. Suitable options are:
\itemize{
\item{"all" for all of the below listed data sets (default),}
\item{"20cr" for 20CR v3,}
\item{"chirps" for CHIRPS v2.0,}
\item{"cmap" for CMAP standard version,}
\item{"cmorph" for CMORPH,}
\item{"cpc" for CPC-Global,}
\item{"cru_ts" for CRU_TS v4.06,}
\item{"em_earth" for EM-EARTH,}
\item{"era20c" for ERA-20C,}
\item{"era5" for ERA5,}
\item{"ghcn" for GHCN-M v2,}
\item{"gldas_clsm" for GLDAS CLSM,}
\item{"gldas_noah" for GLDAS NOAH,}
\item{"gldas_vic" for GLDAS VIC,}
\item{"gpcc" for GPCC v2020,}
\item{"gpcp" for GPCP v2.3,}
\item{"gpm_imerg" for GPM IMERGM Final v06,}
\item{"mswep" for MSWEP v2.8,}
\item{"ncep_doe" for NCEP/DOE,}
\item{"ncep_ncar" for NCEP/NCAR,}
\item{"persiann" for PERSIANN-CDR,}
\item{"precl" for PREC/L,}
\item{"terraclimate" for TerraClimate,}
\item{"trmm_3b43" for TRMM 3B43 v7,}
\item{"udel" for UDEL v501.}
}}

\item{database_path}{a character string with the path where the "database" folder is located.}
}
\value{
No return value, called to subset via cdo
}
\description{
The function \code{mon_to_year} aggregates the requested data set from monthly to yearly time steps and stores it in <project_folder>/data/processed. If the input data set is a Raster object the output will be stored in the same location of the input file.
}
\examples{
\dontrun{
x <- mon_to_year("gpcp", tempdir())
w <- raster::brick("dummie.nc")
z <- mon_to_year(w, tempdir())
}
}
