% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muldpm.R
\name{muldpm}
\alias{muldpm}
\title{Multiply by days per month}
\usage{
muldpm(x)
}
\arguments{
\item{x}{a RasterBrick object with monthly data in [units/day]}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
tavg_brick <- raster::brick('terraclimate_tavg.nc')
pet_od <- pet(method = "od", tavg = tavg_brick)
pet_od <- muldpm(pet_od)
}
}
