% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_check.R
\name{name_check}
\alias{name_check}
\title{Name checker}
\usage{
name_check(data_name)
}
\arguments{
\item{data_name}{a character string.}
}
\value{
No return value, called to download the data set.
}
\description{
Function to check if the data file follows pRecipe naming conventions.
}
\keyword{internal}
