% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_to_csv.R
\name{nc_to_csv}
\alias{nc_to_csv}
\title{Transform raster into data.table}
\usage{
nc_to_csv(data_file)
}
\arguments{
\item{data_file}{a character string with the path to the data file.}
}
\value{
data.table
}
\description{
Function to transform a raster brick into a data.table
}
\examples{
\dontrun{
x <- nc_to_csv("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
y <- nc_to_csv("dummie.nc")
}
}
