% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\title{Nash–Sutcliffe Efficiency}
\usage{
nse(x, ref)
}
\arguments{
\item{x}{a data.table generated by \code{\link{make_ts}}}

\item{ref}{a data.table with data used for evaluation}
}
\value{
numeric
}
\description{
Function for calculating the Nash–Sutcliffe efficiency.
}
