% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_bar}
\alias{plot_bar}
\title{Precipitation bar plot}
\usage{
plot_bar(x)
}
\arguments{
\item{x}{a pRecipe data.table.}
}
\value{
ggplot object
}
\description{
Function for plotting (bar plot) monthly area averaged precipitation.
}
