% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Precipitation box plot}
\usage{
plot_box(x, var = "Precipitation", unit = "mm")
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character, see details)}

\item{var}{character string with the name of the variable to be used for the axis title}

\item{unit}{character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for plotting (boxplot) time-series of area averaged precipitation.
}
\details{
If `x` is a data.table, its columns should be named: "lon", "lat", "date", and "value"

If `x` is a filename, it should point to a *.nc file.

`var` is a character string describing the variable to be used for the plot title

`unit` is a character string describing the unit of measurement to be used for the plot title
}
