% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_line}
\alias{plot_line}
\title{Precipitation line plot}
\usage{
plot_line(x)
}
\arguments{
\item{x}{a pRecipe data.table.}
}
\value{
ggplot object
}
\description{
Function for plotting monthly time-series of area averaged precipitation.
}
