% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_map}
\alias{plot_map}
\title{Precipitation map}
\usage{
plot_map(x, monthly = FALSE)
}
\arguments{
\item{x}{a pRecipe data.table.}

\item{monthly}{logical. If TRUE will generate one plot per month.}
}
\value{
list with ggplot objects
}
\description{
Function for plotting precipitation maps.
}
