% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Precipitation matrix}
\usage{
plot_matrix(x)
}
\arguments{
\item{x}{a pRecipe data.table.}
}
\value{
list with ggplot objects
}
\description{
Function for plotting monthly precipitation matrices.
}
