% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Precipitation summary plot}
\usage{
plot_summary(x, var = "Precipitation", unit = "mm")
}
\arguments{
\item{x}{a csv generated by \code{\link{make_ts}}}

\item{var}{a character string with the name of the variable to be used for the axis title}

\item{unit}{a character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for plotting precipitation summary (line, matrix, box, and density)
}
