% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taylor.R
\name{plot_taylor}
\alias{plot_taylor}
\title{Taylor diagram}
\usage{
plot_taylor(x, y, groups = "default", ...)
}
\arguments{
\item{x}{data.table}

\item{y}{data.table}

\item{groups}{character}

\item{...}{see details}
}
\value{
plot object
}
\description{
Convenient and aesthetic visualization of data in a Taylor diagram.
}
\details{
`x` columns should be named: "lon", "lat", "date", "value", "dataset", and "source".

`y` columns should be named: "lon", "lat", "date", "value", "dataset", and "source".

`groups` character to define panels. Suitable options are:
\itemize{
\item "default" (only one panel)
\item "source"
\item "seasons" (only works properly with monthly data)
}

`...` extra arguments passed on to \code{openair::\link[openair:TaylorDiagram]{TaylorDiagram}}
}
