% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{prate_to_psum}
\alias{prate_to_psum}
\title{Transforms the data from precipitation rate into total precipitation.}
\usage{
prate_to_psum(name, folder_path, overwrite = TRUE)
}
\arguments{
\item{name}{a character string with the name of the desired data set. Suitable options are:
\itemize{
\item{"cmap" for CMAP standard version,}
\item{"cru" for CRU_TS v4.04,}
\item{"gpcp" for GPCP v2.3,}
\item{"gpm" for GPM IMERG Final v06,}
\item{"ncep" for NCEP/NCAR,}
\item{"ncep2" for NCEP/DOE,}
\item{"precl" for PRECL,}
\item{"trmm" for TRMM 3B43 v7,}
\item{"udel" for UDEL v501.}
}}

\item{folder_path}{a character string with the folder path where the data set is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}
}
\description{
The function \code{prate_to_psum} accumulates precipitation of the selected data product.
}
\note{
For Udel data only the units are transformed to mm because it is already in terms of total precipitation
}
