% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{prate_to_psum_gpcp}
\alias{prate_to_psum_gpcp}
\title{GPCP data reader}
\usage{
prate_to_psum_gpcp(file, overwrite)
}
\arguments{
\item{file}{a character string with the path where the data set file is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}
}
\value{
a brick with total monthly precipitation in [mm] at 2.5 degrees for 1979-2019.
}
\description{
Function for reading GPCP NC file.
}
