% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{prate_to_psum_gpm}
\alias{prate_to_psum_gpm}
\title{GPM data reader}
\usage{
prate_to_psum_gpm(file, overwrite, check = FALSE)
}
\arguments{
\item{file}{a character string with the path where the data set file is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}

\item{check}{logical. If TRUE parallel works only with 2 cores for CRAN tests.}
}
\value{
a list of bricks with monthly precipitation rate in [mm/hour] at 0.1 degrees for 2000-2019.
}
\description{
Function for reading GPM HDF5 files.
}
