% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{prate_to_psum_ncep}
\alias{prate_to_psum_ncep}
\title{NCEP/NCAR data reader}
\usage{
prate_to_psum_ncep(file, overwrite)
}
\arguments{
\item{file}{a character string with the path where the data set file is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}
}
\value{
a brick with monthly precipitation rate in [mm/s] at T62 Gaussian grid for 1948-2019.
}
\description{
Function for reading NCEP/NCAR NC files.
}
