% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{prate_to_psum_precl}
\alias{prate_to_psum_precl}
\title{PRECL data reader}
\usage{
prate_to_psum_precl(file, overwrite)
}
\arguments{
\item{file}{a character string with the path where the data set file is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}
}
\value{
a brick with monthly precipitation rate in [mm/day] at 0.5 degrees for 1948-2012.
}
\description{
Function for reading PRECL NC file.
}
