% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{prate_to_psum_trmm}
\alias{prate_to_psum_trmm}
\title{TRMM data reader}
\usage{
prate_to_psum_trmm(file, overwrite, check = FALSE)
}
\arguments{
\item{file}{a character string with the path where the data set file is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}

\item{check}{logical. If TRUE parallel works only with 2 cores for CRAN tests.}
}
\value{
a list of bricks with monthly precipitation rate in [mm/h] at 0.25 degrees for 1998-2019.
}
\description{
Function for reading TRMM 3B43 HDF files.
}
