% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulators.R
\name{prate_to_psum_udel}
\alias{prate_to_psum_udel}
\title{UDEL data reader}
\usage{
prate_to_psum_udel(file, overwrite)
}
\arguments{
\item{file}{a character string with the path where the data set file is located.}

\item{overwrite}{logical. If TRUE (default) the original file with precipitation data will be replaced with a file with total precipitation.}
}
\value{
a brick with total monthly precipitation in [cm] at 0.5 degrees for 1900-2017.
}
\description{
Function for reading UDEL NC file.
}
