% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe_class.R
\docType{class}
\name{pRecipe-class}
\alias{pRecipe-class}
\alias{pRecipe}
\title{An S4 class that extends data.table for pRecipe data}
\description{
An S4 class that extends data.table for pRecipe data
}
\section{Slots}{

\describe{
\item{\code{info}}{character with data set name and type}
}}

\keyword{internal}
