% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\docType{package}
\name{pRecipe-package}
\alias{pRecipe-package}
\title{pRecipe: Precipitation R Recipes}
\description{
An open-access tool/framework to download, validate, visualize, and analyze multi-source precipitation data across various spatio-temporal scales. Ultimately providing the hydrology science community with the tools for consistent and reproducible analysis regarding precipitation.
}
\author{
Mijael Rodrigo Vargas Godoy \email{vargas_godoy@fzp.czu.cz} and Yannis Markonis \email{markonis@fzp.czu.cz}
}
