% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe_masks.R
\docType{data}
\name{pRecipe_masks}
\alias{pRecipe_masks}
\title{Various Masks}
\format{
A data.table with 1036800 obs. of 16 variables:
\describe{
  \item{lon}{(numeric) longitude in degrees}
  \item{lat}{(numeric) latitute in degrees}
  \item{basin_id}{(numeric) Pfafstetter code}
  \item{biome_class}{(character) terrestrial biomes}
  \item{biome_short_class}{(character) terrestrial biomes short names}
  \item{country}{(character) country names}
  \item{country_short}{(character) country alpha-3 code}
  \item{elev_class}{(character) elevation ranges}
  \item{KG_class}{(character) Köppen-Geiger classification}
  \item{KG_class_1}{(character) Köppen-Geiger first classification}
  \item{KG_class_2}{(character) Köppen-Geiger second classification}
  \item{KG_class_3}{(character) Köppen-Geiger third classification}
  \item{KG_class_1_name}{(character) Köppen-Geiger first classification name}
  \item{land_cover_class}{(character) land cover classification}
  \item{land_cover_short_class}{(character) land cover classification short name}
  \item{land_mask}{(character) landmask}
}
}
\usage{
pRecipe_masks
}
\description{
Masks in data.table format
}
\keyword{datasets}
