% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_all}
\alias{reformat_all}
\title{All data downloader}
\usage{
reformat_all(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading GPCP NC file.
}
