% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_cpc}
\alias{reformat_cpc}
\title{CPC data reader}
\usage{
reformat_cpc(folder_path, chk = FALSE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a list of bricks with total daily precipitation in [mm] at 0.5 degrees for 1979-2019.
}
\description{
Function for reading CPC-GLOBAL NC files.
}
