% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{reformat_data}
\alias{reformat_data}
\title{Read precipitation data from various sources and reformat them into .Rds files}
\usage{
reformat_data(folder_path, name)
}
\arguments{
\item{folder_path}{a character string with the folder path where the data set is located.}

\item{name}{a character string with the name of the desired data set. Suitable options are:
\itemize{
\item{"20cr" for 20CR v3,}
\item{"cmap" for CMAP standard version,}
\item{"cpc" for CPC-Global,}
\item{"cru_ts" for CRU_TS v4.04,}
\item{"ghcn" for GHCN-M v2}
\item{"gpcc" for GPCC v2018,}
\item{"gpcp" for GPCP v2.3,}
\item{"gpm_imergm" for GPM IMERGM Final v06,}
\item{"ncep" for NCEP/NCAR,}
\item{"ncep2" for NCEP/DOE,}
\item{"precl" for PRECL,}
\item{"trmm_3b43" for TRMM 3B43 v7,}
\item{"udel" for UDEL v501.}
}}
}
\description{
The function \code{reformat_data} reformats the selected data product.
}
