% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_gpcc}
\alias{reformat_gpcc}
\title{GPCC data reformatter}
\usage{
reformat_gpcc(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\description{
Function for reading GPCC NC file, and reformatting them into data.table which is stored in an .Rds file.
}
