% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_gpm}
\alias{reformat_gpm}
\title{GPM data reader}
\usage{
reformat_gpm(folder_path, save = TRUE, preserve = TRUE, check = FALSE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}

\item{save}{logical. If TRUE (default) an .Rds file will be saved in the same location.}

\item{preserve}{logical. If TRUE (default) the original file will be preserved.}

\item{check}{logical. If TRUE parallel works only with 2 cores for CRAN tests.}
}
\value{
a list of bricks with monthly precipitation rate in [mm/hour] at 0.1 degrees for 2000-2019.
}
\description{
Function for reading GPM HDF5 files.
}
