% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_gpm_imergm}
\alias{reformat_gpm_imergm}
\title{GPM data reformatter}
\usage{
reformat_gpm_imergm(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\description{
Function for reading GPM HDF5 files, and reformatting them into data.table which is stored in an .Rds file.
}
