% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_ncep2}
\alias{reformat_ncep2}
\title{NCEP/DOE data reader}
\usage{
reformat_ncep2(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a brick with monthly precipitation rate in [mm/s] at T62 Gaussian grid for 1979-2020.
}
\description{
Function for reading NCEP/DOE NC files.
}
