% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_ncep_doe}
\alias{reformat_ncep_doe}
\title{NCEP/DOE data reformatter}
\usage{
reformat_ncep_doe(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "raw" folder.}
}
\description{
Function for reading NCEP/DOE NC files, and reformatting them into data.table which is stored in an .Rds file.
}
