% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_ncep_ncar}
\alias{reformat_ncep_ncar}
\title{NCEP/NCAR data reformatter}
\usage{
reformat_ncep_ncar(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\description{
Function for reading NCEP/NCAR NC files, and reformatting them into data.table which is stored in an .Rds file.
}
