% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_trmm}
\alias{reformat_trmm}
\title{TRMM data reader}
\usage{
reformat_trmm(folder_path, chk = FALSE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a list of bricks with monthly precipitation rate in [mm/h] at 0.25 degrees for 1998-2019.
}
\description{
Function for reading TRMM 3B43 HDF files.
}
