% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_udel}
\alias{reformat_udel}
\title{UDEL data reader}
\usage{
reformat_udel(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a brick with total monthly precipitation in [cm] at 0.5 degrees for 1900-2017.
}
\description{
Function for reading UDEL NC file.
}
