% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{Subset a precipitation data product in space}
\usage{
regrid(x, new_res, autosave = FALSE)
}
\arguments{
\item{x}{a character string with the path to the data file. Or a RasterBrick.}

\item{new_res}{numeric. Target resolution must be a multiple of 0.25 (e.g., 0.5, 1, 2.5).}

\item{autosave}{logical FALSE (default). If TRUE data will be automatically stored in the same location of the input file}
}
\value{
An aggregated RasterBrick.
}
\description{
The function \code{regrid} aggregates the requested data sets into desired resolution and stores it in the same location of the input file.
}
\examples{
\dontrun{
x <- regrid("gpcp_tp_mm_global_197901_202205_025_monthly.nc", 1)
z <- regrid("dummie.nc", 1)
}
}
