% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_nc.R
\name{save_nc}
\alias{save_nc}
\title{Save .nc file}
\usage{
save_nc(dummie_nc, nc_out)
}
\arguments{
\item{dummie_nc}{a Raster object.}

\item{nc_out}{a character string.}
}
\value{
No return value, called to save a file
}
\description{
Function to save data compatible with pRecipe in .nc file
}
\keyword{internal}
