% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveNC.R
\name{saveNC}
\alias{saveNC}
\title{Save .nc file}
\usage{
saveNC(x, file)
}
\arguments{
\item{x}{a Raster object.}

\item{file}{a character string.}
}
\value{
No return value, called to save a file
}
\description{
Function to save data compatible with pRecipe in .nc file
}
\examples{
\dontrun{
save_nc(dummie_brick, "gpcp_tp_mm_global_197901_202205_025_monthly.nc")
}
}
