% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sd_20cr}
\alias{sd_20cr}
\title{Download and reformat ensemble spread of 20CR v3 for 1836-1890}
\usage{
sd_20cr(destination)
}
\arguments{
\item{destination}{a character string with the path to the "integration/aux" folder.}
}
\description{
Function for downloading and reformatting 20cr sd values between 1836 and 1890.
}
