% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sellonlatbox.R
\name{sellonlatbox}
\alias{sellonlatbox}
\alias{sellonlatbox,Raster-method}
\alias{sellonlatbox,data.table-method}
\alias{sellonlatbox,character-method}
\title{Select Longitude Latitude Box}
\usage{
sellonlatbox(x, y)

\S4method{sellonlatbox}{Raster}(x, y)

\S4method{sellonlatbox}{data.table}(x, y)

\S4method{sellonlatbox}{character}(x, y)
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character, see details)}

\item{y}{numeric. Bounding box in the form: (xmin, xmax, ymin, ymax)}
}
\value{
Raster* object; data.table
}
\description{
The function \code{sellonlatbox} subsets the data in space within a bounding box.
}
\details{
If x is a data.table, its columns should be named: "lon", "lat", "date", and "value"

If x is a filename, it should point to a *.nc file.
}
\examples{
\dontrun{
download_data("gldas-vic", tempdir(), timestep = "yearly")
r <- raster::brick(paste0(tempdir(),
"/gldas-vic_tp_mm_land_194801_201412_025_yearly.nc"))
s <- sellonlatbox(r, c(12.24, 18.85, 48.56, 51.12))
}
}
