% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{show_info}
\alias{show_info}
\title{Show data content}
\usage{
show_info(x)
}
\arguments{
\item{x}{a character with the path to the desired file}
}
\value{
character vector with screen print out
}
\description{
The function \code{show_info} displays the specification of the desired file.
}
