% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrator.R
\name{split_time}
\alias{split_time}
\title{Data base time splitter}
\usage{
split_time(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "database" folder.}
}
\description{
Function for splitting the database into common time periods.
}
