% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular}
\alias{tabular}
\alias{tabular,Raster-method}
\alias{tabular,character-method}
\title{Transform raster into data.table}
\usage{
tabular(x)

\S4method{tabular}{Raster}(x)

\S4method{tabular}{character}(x)
}
\arguments{
\item{x}{Raster* object; filename (character, see details)}
}
\value{
data.table
}
\description{
Function to transform a raster brick into a data.table
}
\examples{
\dontrun{
download_data("gldas-vic", tempdir(), timestep = "yearly")
r <- raster::brick(paste0(tempdir(),
"/gldas-vic_tp_mm_land_194801_201412_025_yearly.nc"))
s <- tabular(r)
}
}
