% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{trend}
\alias{trend}
\alias{trend,Raster-method}
\alias{trend,data.table-method}
\alias{trend,character-method}
\title{Trends}
\usage{
trend(x)

\S4method{trend}{Raster}(x)

\S4method{trend}{data.table}(x)

\S4method{trend}{character}(x)
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character, see details)}
}
\value{
Raster* object; data.table
}
\description{
The function \code{trend} computes linear slope.
}
\details{
If `x` is a data.table, its columns should be named: "lon", "lat", "date", and "value"

If `x` is a filename, it should point to a *.nc file.
}
