% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_average.R
\name{weighted_average}
\alias{weighted_average}
\title{Weighted average}
\usage{
weighted_average(x)
}
\arguments{
\item{x}{a character string}
}
\value{
data.table variable
}
\description{
Function to compute the weighted average of an .nc file
}
\keyword{internal}
