% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{create_folders}
\alias{create_folders}
\title{Directory creator}
\usage{
create_folders(destination)
}
\arguments{
\item{destination}{a character string with the path where the folders will be located.}
}
\description{
Function for creating tidy data directories.
}
