% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{crop_data}
\alias{crop_data}
\title{Crop precipitation data sets}
\usage{
crop_data(x, shp_path)
}
\arguments{
\item{x}{a pRecipe data.table imported using \code{\link{import_data}}.}

\item{shp_path}{a character string with the path to the ".shp" file.}
}
\value{
a data.table with the cropped data sets
}
\description{
The function \code{crop_data} crops the data sets using a shapefile mask.
}
