% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data.R
\name{crop_data}
\alias{crop_data}
\title{Crop precipitation data sets}
\usage{
crop_data(data_file, shp_path)
}
\arguments{
\item{data_file}{a character string with the path to the data file.}

\item{shp_path}{a character string with the path to the ".shp" file.}
}
\value{
No return value, called to crop and store store the new data file.
}
\description{
The function \code{crop_data} crops the data sets using a shapefile mask and stores it in the same location of the input file.
}
\examples{
\dontrun{
crop_data("gpcp_tp_mm_global_197901_202205_025_monthly.nc", "cze.shp")
crop_data("dummie.nc", "cze.shp")
}
}
