% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_20cr}
\alias{download_20cr}
\title{20CR data downloader}
\usage{
download_20cr(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading 20CR NC files.
}
