% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_all}
\alias{download_all}
\title{All data downloader}
\usage{
download_all(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GPCP NC file.
}
