% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_all}
\alias{download_all}
\title{All data downloader}
\usage{
download_all(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\description{
Function for downloading GPCP NC file.
}
