% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_chirps.R
\name{download_chirps}
\alias{download_chirps}
\title{CHIRPS data downloader}
\usage{
download_chirps(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CHIRPS.
}
\keyword{internal}
