% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_cmap}
\alias{download_cmap}
\title{CMAP data downloader}
\usage{
download_cmap(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CMAP NC files.
}
