% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_cmorph}
\alias{download_cmorph}
\title{CMORPH data downloader}
\usage{
download_cmorph(destination, start_year = 1998, end_year = 2020)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}

\item{start_year}{numeric. Start year should be between 1998-2020.}

\item{end_year}{numeric. End year should be between 1998-2020, and should be greater or equal to start year.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CMORPH NC files.
}
