% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cru_ts.R
\name{download_cru_ts}
\alias{download_cru_ts}
\title{CRU_TS data downloader}
\usage{
download_cru_ts(folder_path = ".", domain = "raw")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}

\item{domain}{a character string with the desired domain data set. Suitable options are:
\itemize{
\item{"raw" for default available spatial coverage,}
\item{"global" for data sets with global (land and ocean) coverage,}
\item{"land" for data sets with land only coverage,}
\item{"ocean", for data sets with ocean only coverage.}
}}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CRU_TS v4.06.
}
\keyword{internal}
