% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_cru_ts}
\alias{download_cru_ts}
\title{CRU data downloader}
\usage{
download_cru_ts(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading CRU_TS NC.GZ file.
}
