% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{download_data}
\alias{download_data}
\title{Download and import precipitation data from various sources}
\usage{
download_data(project_folder, name = "all", reformat = FALSE)
}
\arguments{
\item{project_folder}{a character string with the path where pRecipe will be hosted. Inside it the required subfolders will be created see \code{\link{create_folders}}}

\item{name}{a character string with the name(s) of the desired data set. Suitable options are:
\itemize{
\item{"all" for all of the below listed data sets (default),}
\item{"20cr" for 20CR v3,}
\item{"cmap" for CMAP standard version,}
\item{"cpc" for CPC-Global,}
\item{"cru_ts" for CRU_TS v4.04,}
\item{"ghcn" for GHCN-M v2}
\item{"gpcc" for GPCC v2018,}
\item{"gpcp" for GPCP v2.3,}
\item{"gpm_imergm" for GPM IMERGM Final v06,}
\item{"ncep_ncar" for NCEP/NCAR,}
\item{"ncep_doe" for NCEP/DOE,}
\item{"precl" for PRECL,}
\item{"trmm_3b43" for TRMM 3B43 v7,}
\item{"udel" for UDEL v501.}
}}

\item{reformat}{logical. If TRUE the downloaded datasets are reformatted into data.table and stored in .Rds files. See \code{\link{reformat_data}}}
}
\description{
The function \code{download_data} downloads the selected data product.
}
\examples{
\dontrun{
download_data("~/global_precipitation/pRecipe")
download_data("~/projects/czu/pRecipe", c("cru_ts", "cpc", "ghcn", "gpcp"), reformat = TRUE)
download_data("~/research/pRecipe", c("gpm_imergm", "trmm_3b43"))
}
}
