% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_ghcn}
\alias{download_ghcn}
\title{GHCN-M data downloader}
\usage{
download_ghcn(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GHCN-M NC file.
}
