% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gldas_clsm.R
\name{download_gldas_clsm}
\alias{download_gldas_clsm}
\title{GLDAS_CLSM data downloader}
\usage{
download_gldas_clsm(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GLDAS CLSM.
}
\keyword{internal}
