% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gldas_vic.R
\name{download_gldas_vic}
\alias{download_gldas_vic}
\title{GLDAS_VIC data downloader}
\usage{
download_gldas_vic(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GLDAS VIC.
}
\keyword{internal}
