% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_gpcc}
\alias{download_gpcc}
\title{GPCC data downloader}
\usage{
download_gpcc(folder_path, resolution = 0.5)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}

\item{resolution}{numeric. Data spatial resolution. Suitable options are:
\itemize{
\item{0.5 for 0.5 degree,}
\item{1 for 1 degree,}
\item{2.5 for 2.5 degree.}
}}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GPCC NC file.
}
