% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpcp.R
\name{download_gpcp}
\alias{download_gpcp}
\title{GPCP data downloader}
\usage{
download_gpcp(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GPCP v2.3.
}
\keyword{internal}
