% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpm_imerg.R
\name{download_gpm_imerg}
\alias{download_gpm_imerg}
\title{GPM_IMERG data downloader}
\usage{
download_gpm_imerg(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GPM IMERGM v06.
}
\keyword{internal}
