% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_gpm_imergm}
\alias{download_gpm_imergm}
\title{GPM data downloader}
\usage{
download_gpm_imergm(folder_path, start_year = 2000, end_year = 2020)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}

\item{start_year}{numeric. Start year should be between 2000-2020.}

\item{end_year}{numeric. End year should be between 2000-2020, and should be greater or equal to start year.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading GPM HDF5 files.
}
\note{
user must \href{https://urs.earthdata.nasa.gov}{Create an Earthdata account} and \href{https://disc.gsfc.nasa.gov/earthdata-login}{Link GES DISC}
}
