% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mswep.R
\name{download_mswep}
\alias{download_mswep}
\title{MSWEP data downloader}
\usage{
download_mswep(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading MSWEP v2.8.
}
\keyword{internal}
