% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncep_doe.R
\name{download_ncep_doe}
\alias{download_ncep_doe}
\title{NCEP_DOE data downloader}
\usage{
download_ncep_doe(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading NCEP/DOE.
}
\keyword{internal}
