% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncep_ncar.R
\name{download_ncep_ncar}
\alias{download_ncep_ncar}
\title{NCEP_NCAR data downloader}
\usage{
download_ncep_ncar(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading NCEP/NCAR.
}
\keyword{internal}
