% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_persiann.R
\name{download_persiann}
\alias{download_persiann}
\title{PERSIANN data downloader}
\usage{
download_persiann(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading PERSIANN-CDR.
}
\keyword{internal}
