% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_precl}
\alias{download_precl}
\title{PRECL data downloader}
\usage{
download_precl(destination, resolution)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}

\item{resolution}{numeric. Data spatial resolution. Suitable options are:
\itemize{
\item{0.5 for 0.5 degree,}
\item{1 for 1 degree,}
\item{2.5 for 2.5 degree.}
}}
}
\description{
Function for downloading PRECL NC file.
}
