% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_terraclimate.R
\name{download_terraclimate}
\alias{download_terraclimate}
\title{TERRACLIMATE data downloader}
\usage{
download_terraclimate(folder_path = "./data/raw/")
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading TerraClimate.
}
\keyword{internal}
