% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_trmm}
\alias{download_trmm}
\title{TRMM data downloader}
\usage{
download_trmm(destination, start_year, end_year)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}

\item{start_year}{numeric. Start year should be between 1998-2019.}

\item{end_year}{numeric. End year should be between 1979-2019, and should be greater or equal to start year.}
}
\description{
Function for downloading TRMM 3B43 HDF files from GES DISC.
}
