% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trmm_3b43.R
\name{download_trmm_3b43}
\alias{download_trmm_3b43}
\title{TRMM_3B43 data downloader}
\usage{
download_trmm_3b43(folder_path = ".")
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading TRMM 3B43 v7.
}
\keyword{internal}
