% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{download_udel}
\alias{download_udel}
\title{UDEL data downloader}
\usage{
download_udel(destination)
}
\arguments{
\item{destination}{a character string with the path where the downloaded file is saved.}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading UDEL NC file.
}
