% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_global_ts.R
\docType{data}
\name{gpm_global_ts}
\alias{gpm_global_ts}
\title{Monthly precipitation data}
\format{
A data.table with 256 obs. of 2 variables:
\describe{
  \item{date}{IDate format \%Y-\%m-\%d}
  \item{value}{monthly average values}
}
}
\source{
National Aeronautics and Space Administration (NASA)
}
\usage{
gpm_global_ts
}
\description{
Global GPM-IMERGM monthly precipitation data in mm. More details of the raw data can be found \href{https://gpm.nasa.gov/data/imerg}{here}.
}
\keyword{datasets}
