% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_cmap}
\alias{import_cmap}
\title{CMAP data reader}
\usage{
import_cmap(file_path)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}
}
\description{
Function for reading CMAP NC files.
}
