% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_cpc}
\alias{import_cpc}
\title{CPC data reader}
\usage{
import_cpc(file_path, start_year, end_year)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}

\item{start_year}{numeric. Start year should be between 1979-2019.}

\item{end_year}{numeric. End year should be between 1979-2019, and should be greater or equal to start year.}
}
\description{
Function for reading CPC-GLOBAL NC files.
}
