% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{import_data}
\alias{import_data}
\title{Download and import precipitation data from various sources}
\usage{
import_data(name, folder_path)
}
\arguments{
\item{name}{a character string with the name of the desired data set. Suitable options are:
\itemize{
\item{"20cr" for 20CR v3}
\item{"cmap" for CMAP standard version,}
\item{"cpc" for CPC-Global,}
\item{"cru" for CRU_TS v4.04,}
\item{"ghcn" for GHCN-M v2}
\item{"gpcc" for GPCC v2018,}
\item{"gpcp" for GPCP v2.3,}
\item{"gpm" for GPM IMERG Final v06,}
\item{"ncep" for NCEP/NCAR,}
\item{"ncep2" for NCEP/DOE,}
\item{"precl" for PRECL,}
\item{"trmm" for TRMM 3B43 v7,}
\item{"udel" for UDEL v501.}
}}

\item{folder_path}{a character string with the folder path where the data set is located.}
}
\description{
The function \code{import_data} downloads the selected data product.
}
