% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_ghcn}
\alias{import_ghcn}
\title{GHCN-M data reader}
\usage{
import_ghcn(file_path)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}
}
\description{
Function for reading GHCN-M NC file.
}
