% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_ncep2}
\alias{import_ncep2}
\title{NCEP/DOE data reader}
\usage{
import_ncep2(file_path)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}
}
\description{
Function for reading NCEP/DOE NC files.
}
