% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_trmm}
\alias{import_trmm}
\title{TRMM data reader}
\usage{
import_trmm(file_path, start_year, end_year)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}

\item{start_year}{numeric. Start year should be between 1998-2019.}

\item{end_year}{numeric. End year should be between 1979-2019, and should be greater or equal to start year.}
}
\description{
Function for reading TRMM 3B43 HDF files.
}
\note{
user must \href{"https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+For+an+EarthData+Login+Profile"}{Create an Earthdata account} and \href{https://disc.gsfc.nasa.gov/earthdata-login}{Link GES DISC}
}
