% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import_udel}
\alias{import_udel}
\title{UDEL data reader}
\usage{
import_udel(file_path)
}
\arguments{
\item{file_path}{a character string with the path where the data set file is located.}
}
\description{
Function for reading UDEL NC file.
}
