% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoNC.R
\name{infoNC}
\alias{infoNC}
\alias{infoNC,character-method}
\alias{infoNC,Raster-method}
\title{Show data content}
\usage{
infoNC(x)

\S4method{infoNC}{character}(x)

\S4method{infoNC}{Raster}(x)
}
\description{
The function \code{infoNC} displays the specification of the desired file.
}
