% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ts.R
\name{make_ts}
\alias{make_ts}
\title{Generate time series}
\usage{
make_ts(data_file)
}
\arguments{
\item{data_file}{a character string with the path to the data file.}
}
\value{
No return value, called to generate and store csv time series.
}
\description{
The function \code{make_ts} generates a csv time series and stored in the same location of the input file.
}
\examples{
\dontrun{
make_ts("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
make_ts("dummie.nc")
}
}
