% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{merge_time}
\alias{merge_time}
\title{Data integrator.}
\usage{
merge_time(database_folder_path = "./data/database")
}
\arguments{
\item{database_folder_path}{a character string with the path to the "database" folder.}
<<<<<<< HEAD
=======
}
\value{
No return value, called to merge all available data sets into one.
>>>>>>> dev
}
\description{
Function for merging the available data sets of a given year by weighted average.
}
