% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mon_to_year.R
\name{mon_to_year}
\alias{mon_to_year}
\title{Rescale a precipitation data product in time}
\usage{
mon_to_year(data_file)
}
\arguments{
\item{data_file}{a character string with the path to the data file.}
}
\value{
No return value, called to aggregate and store store the new data file.
}
\description{
The function \code{mon_to_year} aggregates the requested data set from monthly to yearly time steps and stores it in the same location of the input file.
}
\examples{
\dontrun{
mon_to_year("gpcp_tp_mm_global_197901_202205_025_monthly.nc")
mon_to_year("dummie.nc")
}
}
