% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_box}
\alias{plot_box}
\title{Precipitation box plot}
\usage{
plot_box(x)
}
\arguments{
\item{x}{a pRecipe data.table.}
}
\value{
ggplot object
}
\description{
Function for plotting (box plot) monthly area averaged precipitation.
}
