% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Precipitation box plot}
\usage{
plot_box(x, var = "Precipitation", unit = "mm")
}
\arguments{
\item{x}{a data.table generated by \code{\link{fldmean}}}

\item{var}{a character string with the name of the variable to be used for the axis title}

\item{unit}{a character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for plotting (boxplot) time-series of area averaged precipitation.
}
