% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Precipitation box plot}
\usage{
plot_box(dummie)
}
\arguments{
\item{dummie}{a csv generated by \code{\link{make_ts}}}
}
\value{
ggplot object
}
\description{
Function for plotting (boxplot) monthly time-series of area averaged precipitation.
}
