% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Precipitation density}
\usage{
plot_density(dummie)
}
\arguments{
\item{dummie}{a csv generated by \code{\link{make_ts}}}
}
\value{
ggplot object
}
\description{
Function for plotting (density) time-series of area averaged precipitation.
}
