% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Precipitation density}
\usage{
plot_density(x, var = "Precipitation", unit = "mm")
}
\arguments{
\item{x}{a csv generated by \code{\link{fldmean}}}

\item{var}{a character string with the name of the variable to be used for the axis title}

\item{unit}{a character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for plotting (density) time-series of area averaged precipitation.
}
