% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Precipitation heatmap}
\usage{
plot_heatmap(x, unit = "mm")
}
\arguments{
\item{x}{a csv generated by \code{\link{make_ts}}}

\item{unit}{a character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for plotting (heatmap) time-series of area averaged precipitation.
}
