% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Precipitation heatmap}
\usage{
plot_heatmap(dummie)
}
\arguments{
\item{dummie}{a csv generated by \code{\link{make_ts}}}
}
\value{
ggplot object
}
\description{
Function for plotting (heatmap) time-series of area averaged precipitation.
}
