% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{Precipitation line plot}
\usage{
plot_line(x, var = "Precipitation", unit = "mm")
}
\arguments{
\item{x}{a csv generated by \code{\link{make_ts}}}

\item{var}{a character string with the name of the variable to be used for the axis title}

\item{unit}{a character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for plotting (line) time-series of area averaged precipitation.
}
