% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{Precipitation line plot}
\usage{
plot_line(dummie)
}
\arguments{
\item{dummie}{a csv generated by \code{\link{make_ts}}}
}
\value{
ggplot object
}
\description{
Function for plotting (line) monthly time-series of area averaged precipitation.
}
