% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Precipitation map plot}
\usage{
plot_map(x, unit = "mm")
}
\arguments{
\item{x}{a .nc file with precipitation}

\item{unit}{a character string with the variable's unit of measurement to be used for the axis title}
}
\value{
ggplot object
}
\description{
Function for mapping the first layer of a .nc file
}
