% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taylor.R
\name{plot_taylor}
\alias{plot_taylor}
\title{Precipitation taylor diagram}
\usage{
plot_taylor(dummie, ref, groups = "type")
}
\arguments{
\item{dummie}{a csv generated by \code{\link{make_ts}}}

\item{ref}{a csv to be used as reference in the same format as that generated by \code{\link{make_ts}}. I.e., with columns: "date", "value", "name", and "type".}

\item{groups}{a string character to define panels. Suitable options are:
\itemize{
\item "type" (default)
\item "seasons" (only works properly with monthly data)
}}
}
\value{
plot object
}
\description{
Function for plotting (taylor diagram) time-series of area averaged precipitation
}
