% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\docType{package}
\name{pRecipe-package}
\alias{pRecipe-package}
\title{pRecipe: Precipitation R Recipes}
\description{
pRecipe is an R package that facilitates the download, exploration, visualization, and analysis of multi-source precipitation data.
}
\references{
Vargas Godoy, M. R., & Markonis, Y. (2023). pRecipe: A global precipitation climatology toolbox and database. Environmental Modelling & Software, 165, 105711. \doi{10.1016/j.envsoft.2023.105711}
}
\author{
\strong{Coded by:} Mijael Rodrigo Vargas Godoy \email{vargas_godoy@fzp.czu.cz}

\strong{Conceptual design by:} Mijael Rodrigo Vargas Godoy \email{vargas_godoy@fzp.czu.cz} and Yannis Markonis \email{markonis@fzp.czu.cz}
}
