% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{crop_data}
\alias{csi}
\alias{far}
\alias{fldmean}
\alias{infoNC}
\alias{muldpm}
\alias{nse}
\alias{pRecipe_masks}
\alias{pod}
\alias{remap}
\alias{saveNC}
\alias{subset_data}
\alias{tabular}
\alias{trend}
\alias{yearstat}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{twc}{\code{\link[twc]{crop_data}}, \code{\link[twc]{csi}}, \code{\link[twc]{far}}, \code{\link[twc]{fldmean}}, \code{\link[twc]{infoNC}}, \code{\link[twc]{muldpm}}, \code{\link[twc]{nse}}, \code{\link[twc]{pod}}, \code{\link[twc]{pRecipe_masks}}, \code{\link[twc]{remap}}, \code{\link[twc]{saveNC}}, \code{\link[twc]{subset_data}}, \code{\link[twc]{tabular}}, \code{\link[twc]{trend}}, \code{\link[twc]{yearstat}}}
}}

