% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_cmap}
\alias{reformat_cmap}
\title{CMAP data reformatter}
\usage{
reformat_cmap(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the "raw" folder is located.}
}
\description{
Function for reading CMAP NC files, and reformatting them into data.table which is stored in an .Rds file.
}
