% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_cmap}
\alias{reformat_cmap}
\title{CMAP data reader}
\usage{
reformat_cmap(folder_path, save = TRUE, preserve = TRUE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}

\item{save}{logical. If TRUE (default) an .Rds file will be saved in the same location.}

\item{preserve}{logical. If TRUE (default) the original file will be preserved.}
}
\value{
a brick with monthly precipitation rate in [mm/day] at 2.5 degrees for 1979-2019.
}
\description{
Function for reading CMAP NC files.
}
