% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{reformat_data}
\alias{reformat_data}
\title{Download and import precipitation data from various sources}
\usage{
reformat_data(name, folder_path, save = TRUE, preserve = TRUE)
}
\arguments{
\item{name}{a character string with the name of the desired data set. Suitable options are:
\itemize{
\item{"20cr" for 20CR v3}
\item{"cmap" for CMAP standard version,}
\item{"cpc" for CPC-Global,}
\item{"cru" for CRU_TS v4.04,}
\item{"ghcn" for GHCN-M v2}
\item{"gpcc" for GPCC v2018,}
\item{"gpcp" for GPCP v2.3,}
\item{"gpm" for GPM IMERG Final v06,}
\item{"ncep" for NCEP/NCAR,}
\item{"ncep2" for NCEP/DOE,}
\item{"precl" for PRECL,}
\item{"trmm" for TRMM 3B43 v7,}
\item{"udel" for UDEL v501.}
}}

\item{folder_path}{a character string with the folder path where the data set is located.}

\item{save}{logical. If TRUE (default) an .Rds file will be saved in the same location.}

\item{preserve}{logical. If TRUE (default) the original file will be preserved.}
}
\description{
The function \code{reformat_data} downloads the selected data product.
}
