% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_ghcn}
\alias{reformat_ghcn}
\title{GHCN-M data reader}
\usage{
reformat_ghcn(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a brick with total monthly precipitation in [mm] at 5 degrees for 1900-2015.
}
\description{
Function for reading GHCN-M NC file.
}
