% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_ghcn}
\alias{reformat_ghcn}
\title{GHCN-M data reformatter}
\usage{
reformat_ghcn(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "raw" folder.}
}
\description{
Function for reading GHCN-M NC file, and reformatting them into data.table which is stored in an .Rds file.
}
