% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_gpcc}
\alias{reformat_gpcc}
\title{GPCC data reader}
\usage{
reformat_gpcc(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a brick with total monthly precipitation in [mm] at 0.5 degrees for 1891-2016.
}
\description{
Function for reading GPCC NC file.
}
