% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_gpm}
\alias{reformat_gpm}
\title{GPM data reader}
\usage{
reformat_gpm(folder_path, chk = FALSE)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}
}
\value{
a list of bricks with monthly precipitation rate in [mm/hour] at 0.1 degrees for 2000-2019.
}
\description{
Function for reading GPM HDF5 files.
}
