% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_ncep_doe}
\alias{reformat_ncep_doe}
\title{NCEP/DOE data reader}
\usage{
reformat_ncep_doe(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path where the data set file is located.}

\item{save}{logical. If TRUE (default) an .Rds file will be saved in the same location.}

\item{preserve}{logical. If TRUE (default) the original file will be preserved.}
}
\value{
a brick with monthly precipitation rate in [mm/s] at T62 Gaussian grid for 1979-2019.
}
\description{
Function for reading NCEP/DOE NC files.
}
