% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatters.R
\name{reformat_trmm_3b43}
\alias{reformat_trmm_3b43}
\title{TRMM data reformatter}
\usage{
reformat_trmm_3b43(folder_path)
}
\arguments{
\item{folder_path}{a character string with the path to the "raw" folder.}
}
\description{
Function for reading TRMM 3B43 HDF files, and reformatting them into data.table which is stored in an .Rds file.
}
