% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{Spatial aggregation}
\usage{
regrid(x, new_res, autosave = FALSE)
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character; see details)}

\item{res}{numeric. Target resolution must be a multiple of 0.25 (e.g., 0.5, 1, 2.5).}
}
\value{
Raster* object; data.table
}
\description{
The function \code{regrid} aggregates data into a new grid resolution.
}
\details{
If x is a data.table, its columns should be named: "lon", "lat", "date", and "value"

If x is a filename, it should point to a *.nc file.
}
\examples{
\dontrun{
x <- regrid("gpcp_tp_mm_global_197901_202205_025_monthly.nc", 1)
z <- regrid("dummie.nc", 1)
}
}
