% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selyear.R
\name{selyear}
\alias{selyear}
\title{Subset a precipitation data product in time}
\usage{
selyear(x, years, autosave = FALSE)
}
\arguments{
\item{x}{a character string with the path to the data file. Or a RasterBrick.}

\item{years}{numeric vector. Time range in the form: (start_year, end_year)}

\item{autosave}{logical FALSE (default). If TRUE data will be automatically stored in the same location of the input file}
}
\value{
A subsetted RasterBrick.
}
\description{
The function \code{selyear} subsets in time the requested data set and stores it in the same location of the input file.
}
\examples{
\dontrun{
selyear("gpcp_tp_mm_global_197901_202205_025_monthly.nc", 
c(2000, 2010), autosave = TRUE)
selyear("dummie.nc", c(2000, 2010), autosave = TRUE)
}
}
