% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{subset_data}
\alias{subset_data}
\title{Subset precipitation data sets}
\usage{
subset_data(data_list, start_year, end_year, country)
}
\arguments{
\item{data_list}{a list generated by \code{select_data}}

\item{start_year}{numeric.}

\item{end_year}{numeric.}

\item{country}{shp file.}
}
\value{
a list with the subsetted data sets
}
\description{
The function \code{subset_data} imports the requested data sets.
}
