% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRecipe.R
\name{subset_data}
\alias{subset_data}
\title{Subset precipitation data sets}
\usage{
subset_data(x, start_year, end_year, box)
}
\arguments{
\item{x}{a pRecipe data.table imported using \code{import_data}.}

\item{start_year}{numeric.}

\item{end_year}{numeric.}

\item{box}{numeric vector. Bounding box in the form: (xmin, ymin, xmax, ymax).}
}
\value{
a data.table with the subsetted data sets
}
\description{
The function \code{subset_data} subsets the imported data sets.
}
\examples{
\donttest{
x <- import_data("~/projects/czu/pRecipe/data/database", c("cru_ts", "cpc", "ghcn", "gpcp"))
y <- subset_data(x, 2000, 2009, c(12.24, 48.56, 18.85, 51.12))
}
}
