% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_space.R
\name{subset_space}
\alias{subset_space}
\title{Subset a precipitation data product in space}
\usage{
subset_space(data_file, bbox)
}
\arguments{
\item{data_file}{a character string with the path to the data file.}

\item{bbox}{numeric vector. Bounding box in the form: (xmin, xmax, ymin, ymax).}
}
\value{
No return value, called to subset and store store the new data file.
}
\description{
The function \code{subset_space} subsets (space) the requested data set and stores it in the same location of the input file.
}
\examples{
\dontrun{
subset_space("gpcp_tp_mm_global_197901_202205_025_monthly.nc",
c(12.24, 18.85, 48.56, 51.12))
subset_space("dummie.nc", c(12.24, 18.85, 48.56, 51.12))
}
}
